ts-chart: extract TimeSpans module
